requirements.txt
streamlit
PyPDF2
python-docx
openpyxl
python-pptx
huggingface_hub
transformers
torch